Move DocumentVisiblityState into compat/ folder
