+++ packages/liveblocks-core/src/compat/DocumentVisibilityState.ts
